This works for me. I see some improvement in both query types, but not huge improvements.

Before:
```shell
$ python run_benchmark.py --search vector --limit 1000
Finished retrieving 1000 vector search query results
Ran search in: 20.7342 sec
```
That's 50 QPS

After (with 20 thread `ThreadPoolExecutor`):
```shell
$ python run_benchmark.py --search vector --limit 1000
Finished retrieving 1000 vector search query results
Ran search in: 14.6384 sec
```

That's 68 qps.

When I run with `py-spy`, the bottleneck seems to be with the `SentenceTransformer` model, which seems to hold the GIL. But I think you are getting the same performance issue with ElasticSearch.

![Screenshot 2023-11-14 at 10 55 04 AM](https://github.com/prrao87/lancedb-study/assets/5488879/8eabdf8a-cf5f-4e98-8919-c3131039c223)