@wjones127 2023-11-14
When I run your benchmark for elasticsearch, it's actually slower than the LanceDB version, even without async:

```
$ python run_benchmark.py --search vector --limit 1000
Finished retrieving 1000 vector search query results
Ran search in: 23.6893 sec
```

That's 43 QPS.

But again I think the bottleneck is the embedding not the retrieval.