@prrao87 2023-11-14
@wjones127 this is exactly the result I was expecting to get all along -- I think it makes perfect sense that each API call would involve a CPU-bound component (to transform the text into an embedding) and an I/O-bound component (due to data transfer between the storage layer and the client), and so it's not possible to truly be free of the GIL in this case. I agree that the embedding will always be the bottleneck when compared to retrieval -- it's just the fact that I was seeing a linear increase in response time with lancedb (but not Elastic) didn't seem right to me. This multi-threaded approach makes perfect sense to speed things up over the purely sequential version I had before.

This works well, thanks a ton for looking into this!